Concluindo e testando regras de negócio de autores e usando Lombok para atribuir construtores

+++ libraryapi/src/main/java/io/github/henrique0120/libraryapi/service/AutorService.java
@@ -1,9 +1,13 @@
 package io.github.henrique0120.libraryapi.service;
 
 import io.github.henrique0120.libraryapi.controller.dto.AutorDTO;
+import io.github.henrique0120.libraryapi.exceptions.LivroAutorException;
+import io.github.henrique0120.libraryapi.exceptions.RegistroDuplicadoException;
 import io.github.henrique0120.libraryapi.model.Autor;
 import io.github.henrique0120.libraryapi.repository.AutorRepository;
+import io.github.henrique0120.libraryapi.repository.LivroRepository;
 import io.github.henrique0120.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -12,15 +16,12 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository repository;
     private final AutorValidator validator;
-
-    public AutorService(AutorRepository repository, AutorValidator validator){
-        this.repository = repository;
-        this.validator = validator;
-    }
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor){
         validator.validar(autor);
@@ -40,6 +41,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if(possuiLivro(autor)){
+            throw new LivroAutorException("Não é permitido excluir um Autor que possui Ator possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -52,6 +56,10 @@ public class AutorService {
             return repository.findByNacionalidade(nacionalidade);
         }
         return repository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 
 }
